chore: moved router codes
